Refactor trace writer atomic helper
